jd_optimizer/requirements.txt
Flask==3.0.2
google-generativeai>=0.7.0
python-dotenv>=1.0.0
gunicorn==21.2.0
